\name{A4}
\alias{A4}
\title{A4}
\description{Internal function for the SIM package used in \code{sim.plot.zscore.heatmap}.}
\usage{
A4(filename, landscape = T)
}

\arguments{
  \item{filename}{ Describe \code{filename} here }
  \item{landscape}{ Describe \code{landscape} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}

\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\examples{
#do not run
}
\keyword{internal}

