\name{chrom.table}
\alias{chrom.table}
\docType{data}
\title{Table with chromosome arm positions}
\description{
   A table indicating the base positions of the beginning and end of chromosome arms. 
   Currently based on the UCSC March 2006 / NCBI 36 build of the human genome.
}
\usage{data(chrom.table)}
\format{
  A data frame with 48 observations on the following 6 variables.
  \describe{describe}
}
\details{
   If necessary, more details than the description above 
}
\source{
   reference to a publication or URL from which the data were obtained 
}
\references{
   possibly secondary sources and usages 
}
\examples{
data(chrom.table)
## maybe str(chrom.table) ; plot(chrom.table) ...
}
\keyword{datasets}
