\name{collapse.arms}
\alias{collapse.arms}
\title{collapse.arms}
\description{
   Internal function for the SIM package used by \code{sim.update.chrom.table}. 
}
\usage{
collapse.arms(chr.bands)
}

\arguments{
  \item{chr.bands}{ Describe \code{chr.bands} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\note{ further notes 

 Make other sections like Warning with \\section\{Warning\}\{...\} 
}
\seealso{ objects to See Also as \code{\link{help}}}
\examples{
#
}
\keyword{internal}

