\name{create.color.vector}
\alias{create.color.vector}
\title{create.color.vector}
\description{
   Internal function for the SIM package used by the \code{sim.plot.zscore.heatmap} function. 
}
\usage{
create.color.vector(my.class, levels = NULL, color.vector = NULL, my.start = 0, my.end = 0.7)
}

\arguments{
  \item{my.class}{ Describe \code{my.class} here }
  \item{levels}{ Describe \code{levels} here }
  \item{color.vector}{ Describe \code{color.vector} here }
  \item{my.start}{ Describe \code{my.start} here }
  \item{my.end}{ Describe \code{my.end} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Melle Sieswerda, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\note{ further notes 

 Make other sections like Warning with \section{Warning }{....} 
}
\seealso{ objects to See Also as \code{\link{sim.plot.zscore.heatmap}}}
\examples{
#
}
\keyword{internal}

