\name{expr.data}
\alias{expr.data}
\docType{data}
\title{Expression data example}
\description{
   Expression log ratios derived from Pollack et al. 2002 PNAS 99:12963-8. 
}
\usage{data(expr.data)}
\format{
  A data frame with 4865 observations on the following 45 variables.
  \describe{describe}
}
\details{
   If necessary, more details than the description above 
}
\source{
   reference to a publication or URL from which the data were obtained 
}
\references{
   possibly secondary sources and usages 
}
\examples{
data(expr.data)
## maybe str(expr.data) ; plot(expr.data) ...
}
\keyword{datasets}
