\name{get.indep.zscores}
\alias{get.indep.zscores}
\title{A function to calculate mean associations}
\description{
   Internal function for the SIM package used to calculate mean z-scores 
   for the independent features across the significant dependent features. This 
   function is called by sim.plot.zscore.heatmap and tabulate.top.indep.features.
}
\usage{
get.indep.zscores(arg, run.name)
}

\arguments{
  \item{arg}{ Describe \code{arg} here }
  \item{run.name}{ Describe \code{run.name} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\seealso{\code{\link{sim.plot.zscore.heatmap}},\code{\link{tabulate.top.indep.features}} }
\examples{
##do not run
}
\keyword{internal}

