\name{get.pvals}
\alias{get.pvals}
\title{get.pvals}
\description{
   Internal function for the SIM package used in all functions working with the 
   calculated p-values. 
}
\usage{
get.pvals(input.regions, run.name)
}

\arguments{
  \item{input.regions}{ Describe \code{input.regions} here }
  \item{run.name}{ Describe \code{run.name} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Melle Sieswerda, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\note{ further notes 

 Make other sections like Warning with \section{Warning }{....} 
}
\seealso{\code{\link{sim.plot.pvals.on.region}},\code{\link{sim.plot.pvals.on.genome}}, 
\code{\link{sim.plot.adj.pvals}}, \code{\link{sim.plot.zscore.heatmap}}, 
\code{\link{tabulate.pvals}}, \code{\link{tabulate.top.dep.features}}, 
\code{\link{tabulate.top.indep.features}}  
}
\examples{
#
}
\keyword{internal}

