\name{get.start.input.regions}
\alias{get.start.input.regions}

\title{get.start.input.regions}
\description{
   Internal function for the SIM package used in the \code{sim.plot.pvals.on.region} function. 
}
\usage{
get.start.input.regions(input.region, run.name)
}
\arguments{
  \item{input.region}{ Describe \code{input.region} here }
  \item{run.name}{ Describe \code{run.name} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Melle Sieswerda, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\note{ further notes 

 Make other sections like Warning with \\section\{Warning\}\{...\} 
}
\seealso{ objects to See Also as \code{\link{help}}}
\examples{
#
}
\keyword{internal}

