\name{run.ia}
\alias{run.ia}

\title{ function to do ...  }
\description{
   Internal function for the SIM package, used by the \code{integrated.analysis} function. 
}
\usage{
run.ia(indep.chr.matrix, dep.chr.matrix, subtype, zscores, method = c("auto", "asymptotic", "permutations", "gamma"))
}

\arguments{
  \item{indep.chr.matrix}{ Describe \code{indep.chr.matrix} here }
  \item{dep.chr.matrix}{ Describe \code{dep.chr.matrix} here }
  \item{subtype}{ Describe \code{subtype} here }
  \item{zscores}{ Describe \code{zscores} here }
  \item{method}{ Describe \code{method} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\seealso{\code{\link{integrated.analysis}},\code{\link[globaltest]{globaltest}}}
\examples{
#
}
\keyword{internal}

