\name{samples}
\alias{samples}
\docType{data}
\title{Samples for example data}
\description{
   Vector of sample names corresponding to the column headers containing the 
   data in both the copy number (\code{acgh.data}) and expression 
   (\code{expr.data}) example datasets. 
}
\usage{data(samples)}
\format{
  The format is:
 chr [1:41] "BT474" "MCF7" "NORWAY.10" "NORWAY.100" "NORWAY.101" ...
}
\details{
   If necessary, more details than the description above 
}
\source{
   reference to a publication or URL from which the data were obtained 
}
\references{
   possibly secondary sources and usages 
}
\examples{
data(samples)
## maybe str(samples) ; plot(samples) ...
}
\keyword{datasets}
