\name{sim.plot.adj.pvals}
\alias{sim.plot.adj.pvals}

\title{sim.plot.adj.pvals}
\description{
   Internal function for the SIM package used for plotting multiple testing 
   corrected p-values in the \code{sim.plot.pvals.on.region} function. 
}
\usage{
sim.plot.adj.pvals(input.region.name, adjust.method = c("BY", "BH", "raw"), run.name = NULL)
}

\arguments{
  \item{input.region.name}{ Describe \code{input.region.name} here }
  \item{adjust.method}{ Describe \code{adjust.method} here }
  \item{run.name}{ Describe \code{run.name} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ put references to the literature/web site here  }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}
\note{ further notes 

 Make other sections like Warning with \\section\{Warning\}\{...\} 
}
\seealso{ objects to See Also as \code{\link{help}}}
\examples{
#
}
\keyword{internal}

