\name{sim.update.chrom.table}
\alias{sim.update.chrom.table}

\title{sim.update.chrom.table}
\description{
   A function to update the genomic positions of chromosome arms. Base locations 
   of the start and end of chromosome arms should be used from the same 
   organism and build of genome as the location provided as annotation with 
   the datasets.
}
\usage{
sim.update.chrom.table(db = "homo_sapiens_core_40_36b")
}

\arguments{
  \item{db}{ Describe \code{db} here }
}
\details{
   If necessary, more details than the description above 
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{ www.ensembl.org  }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}

\seealso{ objects to See Also as \code{\link{help}}}
\examples{
#
}
\keyword{internal}

