\name{writeln}
\alias{writeln}

\title{writeln}
\description{
     Internal function for the SIM package, used by all main functions that 
     print an action to the screen.   
}
\usage{
writeln(txt = "", ...)
}

\arguments{
  \item{txt}{   Describe \code{txt} here   }
  \item{\dots}{   Describe \code{\dots} here   }
}
\details{
     If necessary, more details than the description above   
}
\value{
   Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{  put references to the literature/web site here   }
\author{Marten Boetzer, Renee X. de Menezes  \email{R.X.Menezes@lumc.nl}}
\note{   further notes   

  Make other sections like Warning with \section{Warning }{....} 
}
\seealso{   objects to See Also as \code{\link{help}} }
\examples{
#
}

\keyword{internal}

