\name{acgh.data}
\alias{acgh.data}
\docType{data}
\title{Array-comparative genomic hybridization data example}
\description{Copy number log ratios derived from Pollack et al. 2002 PNAS 99:12963-8.}
\usage{data(acgh.data)}
\format{
  A data frame with 4865 observations on the following 45 variables.
  \describe{describe}
}
\details{
   If necessary, more details than the description above 
}
\source{
   reference to a publication or URL from which the data were obtained 
}
\references{
   possibly secondary sources and usages 
}
\examples{
data(acgh.data)
## maybe str(acgh.data) ; plot(acgh.data) ...
}
\keyword{datasets}
