\name{samples}

\alias{samples}

\docType{data}

\title{Samples for example data}

\description{
	Vector of sample names corresponding to the column headers containing the 
   	data in both the copy number (\code{acgh.data}) and expression (\code{expr.data}) example datasets. 
}

\usage{
	data(samples)
}

\format{
  	A character vector.
}

\source{
	\url{http://genome-www.stanford.edu/aCGH_breast/data.shtml} 
}

\references{
Pollack JR, Sorlie T, Perou CM, Rees CA, Jeffrey SS, Lonning PE, Tibshirani R, Botstein D, Borresen-Dale AL, Brown PO (2002).
Microarray analysis reveals a major direct role of DNA copy number alteration in the transcriptional program of human breast tumors.
\emph{Proc Natl Acad Sci USA}. \bold{1}, 99(20), 12963-8. 	  
}

\examples{
	data(samples)
}

\keyword{datasets}
